+++ src/views/home/home.js
@@ -1,72 +1,21 @@
 import { main } from '/layouts/main.js';
 await main();
 import { getItemData, getItemByCategory } from '/lib/Fetcher.js';
+import { createItem } from './create-item.js';
 
 const amountAll = document.querySelector('.prod__item--amount');
+const list = document.querySelector('.prod__list');
 
-//item 리스트 컴포넌트 만들기
-const createItem = (item) => {
-  const itemElement = document.createElement('li');
-  itemElement.classList.add('prod__item');
-
-  const link = document.createElement('a');
-  link.classList.add('prod__link');
-  link.setAttribute('href', `/products?id=${item.id}`);
-
-  const thumb = document.createElement('img');
-  thumb.classList.add('prod__link-thumb');
-  thumb.setAttribute('src', item.imgUrl);
-  thumb.setAttribute('alt', `${item.name} 대표 이미지`);
-  //이미지 로드 실패 시 디폴트 이미지
-  thumb.onerror = () => {
-    thumb.onerror = null;
-    thumb.setAttribute('src', '../img/error.png');
-    thumb.style.border = '1px solid #e8e8e8';
-  };
-
-  const info = document.createElement('div');
-  info.classList.add('prod__info');
-
-  const title = document.createElement('p');
-  title.classList.add('prod__title');
-  title.innerText = item.name;
-
-  const order = document.createElement('div');
-  order.classList.add('prod__order');
-
-  const price = document.createElement('span');
-  price.innerHTML = `<strong class="prod__order-price">${item.price.toLocaleString()}</strong>원`;
-
-  order.appendChild(price);
-  info.appendChild(title);
-  info.appendChild(order);
-  link.appendChild(thumb);
-  link.appendChild(info);
-  itemElement.appendChild(link);
-
-  return itemElement;
-};
-
+// 전체 아이템 렌더링
 const items = await getItemData();
 amountAll.innerText = items.length;
-const list = document.querySelector('.prod__list');
 
 items.forEach((item) => {
   const itemElement = createItem(item);
   list.appendChild(itemElement);
 });
 
-const category = document.querySelectorAll('.nav__cate li');
-category.forEach((cate) => {
-  cate.addEventListener('click', (e) => {
-    //기존 on카테고리에서 on클래스 삭제하고
-    document.querySelector('.nav__cate--on').classList.remove('nav__cate--on');
-    //클릭한 카테고리에 on 클래스 추가
-    e.target.classList.add('nav__cate--on');
-  });
-});
-
-//카테고리별 아이템 렌더링
+// 카테고리별 아이템 렌더링
 const categories = document.querySelectorAll('.nav__cate li');
 
 categories.forEach((category) => {
@@ -84,13 +33,13 @@ categories.forEach((category) => {
       window.location.href = '/';
       return;
     } else {
-      getCategoryItem(selectedCategory);
+      renderCategoryItem(selectedCategory);
     }
   };
   category.addEventListener('click', renderProducts);
 });
 
-const getCategoryItem = async (selectedCategory) => {
+const renderCategoryItem = async (selectedCategory) => {
   const CategoryItems = await getItemByCategory(selectedCategory);
   amountAll.innerText = CategoryItems.length;
   const list = document.querySelector('.prod__list');
@@ -99,11 +48,4 @@ const getCategoryItem = async (selectedCategory) => {
     const itemElement = createItem(item);
     list.appendChild(itemElement);
   });
-
-  // URL 변경 코드
-  const currentUrl = window.location.href;
-  // /products/뒤에오는 문자열 찾기-> 카테고리명으로 변경하기
-  const newUrl = currentUrl.replace(/\/products\/(.*)\/?/, `/products/${selectedCategory}`);
-  // 브라우저 히스토리에 새 url추가
-  window.history.pushState({ path: newUrl }, '', newUrl);
 };
